game/requirements.txt
azure-identity
azure-storage-blob
flask
opencv-python
Pillow
python-dotenv
json